push-service/requirements.txt
celery==5.5.3
fastapi==0.120.0
fastapi[standard]
firebase_admin==7.1.0
python-dotenv==1.2.1
requests==2.32.5
uvicorn==0.35.0

pydantic~=2.12.3
certifi~=2025.10.5
firebase_admin~=7.1.0
firebase_admin~=7.1.0
Jinja2~=3.1.6
pika~=1.3.2
firebase_admin~=7.1.0
kombu~=5.5.4
firebase_admin~=7.1.0